Release

+++ packages/plugins/knowledge-base/src/help-center-article/help-center-article.service.ts
@@ -1,11 +1,12 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { CrudService } from '@gauzy/core';
+import { TenantAwareCrudService } from '@gauzy/core';
+import { isNotEmpty } from '@gauzy/common';
 import { HelpCenterArticle } from './help-center-article.entity';
 
 @Injectable()
-export class HelpCenterArticleService extends CrudService<HelpCenterArticle> {
+export class HelpCenterArticleService extends TenantAwareCrudService<HelpCenterArticle> {
 	constructor(
 		@InjectRepository(HelpCenterArticle)
 		private readonly helpCenterArticle: Repository<HelpCenterArticle>
@@ -25,6 +26,8 @@ export class HelpCenterArticleService extends CrudService<HelpCenterArticle> {
 	}
 
 	async deleteBulkByCategoryId(ids: string[]) {
-		return await this.repository.delete(ids);
+		if (isNotEmpty(ids)) {
+			return await this.repository.delete(ids);
+		}
 	}
 }
